new openapi spec
