+++ stream-video-android-core/src/main/baseline-prof.txt
@@ -972,15 +972,15 @@ HSPLio/getstream/video/android/core/internal/module/ConnectionModule$devicesApi$
 PLio/getstream/video/android/core/internal/module/ConnectionModule$devicesApi$2;->invoke()Ljava/lang/Object;
 HSPLio/getstream/video/android/core/internal/module/ConnectionModule$devicesApi$2;->invoke()Lorg/openapitools/client/apis/DevicesApi;
 PLio/getstream/video/android/core/internal/module/ConnectionModule$devicesApi$2;->invoke()Lorg/openapitools/client/apis/DevicesApi;
-Lio/getstream/video/android/core/internal/module/ConnectionModule$eventsApi$2;
-HSPLio/getstream/video/android/core/internal/module/ConnectionModule$eventsApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
-PLio/getstream/video/android/core/internal/module/ConnectionModule$eventsApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+Lio/getstream/video/android/core/internal/module/connectionModule.api$2;
+HSPLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+PLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
 Lio/getstream/video/android/core/internal/module/ConnectionModule$livestreamingApi$2;
 HSPLio/getstream/video/android/core/internal/module/ConnectionModule$livestreamingApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
 PLio/getstream/video/android/core/internal/module/ConnectionModule$livestreamingApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
-Lio/getstream/video/android/core/internal/module/ConnectionModule$moderationApi$2;
-HSPLio/getstream/video/android/core/internal/module/ConnectionModule$moderationApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
-PLio/getstream/video/android/core/internal/module/ConnectionModule$moderationApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+Lio/getstream/video/android/core/internal/module/connectionModule.api$2;
+HSPLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+PLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
 Lio/getstream/video/android/core/internal/module/ConnectionModule$networkStateProvider$2;
 HSPLio/getstream/video/android/core/internal/module/ConnectionModule$networkStateProvider$2;-><init>(Landroid/content/Context;)V
 PLio/getstream/video/android/core/internal/module/ConnectionModule$networkStateProvider$2;-><init>(Landroid/content/Context;)V
@@ -1005,11 +1005,11 @@ HSPLio/getstream/video/android/core/internal/module/ConnectionModule$retrofit$2;
 PLio/getstream/video/android/core/internal/module/ConnectionModule$retrofit$2;->invoke()Ljava/lang/Object;
 HSPLio/getstream/video/android/core/internal/module/ConnectionModule$retrofit$2;->invoke()Lretrofit2/Retrofit;
 PLio/getstream/video/android/core/internal/module/ConnectionModule$retrofit$2;->invoke()Lretrofit2/Retrofit;
-Lio/getstream/video/android/core/internal/module/ConnectionModule$videoCallsApi$2;
-HSPLio/getstream/video/android/core/internal/module/ConnectionModule$videoCallsApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
-PLio/getstream/video/android/core/internal/module/ConnectionModule$videoCallsApi$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
-PLio/getstream/video/android/core/internal/module/ConnectionModule$videoCallsApi$2;->invoke()Ljava/lang/Object;
-PLio/getstream/video/android/core/internal/module/ConnectionModule$videoCallsApi$2;->invoke()Lorg/openapitools/client/apis/VideoCallsApi;
+Lio/getstream/video/android/core/internal/module/connectionModule.api$2;
+HSPLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+PLio/getstream/video/android/core/internal/module/connectionModule.api$2;-><init>(Lio/getstream/video/android/core/internal/module/ConnectionModule;)V
+PLio/getstream/video/android/core/internal/module/connectionModule.api$2;->invoke()Ljava/lang/Object;
+PLio/getstream/video/android/core/internal/module/connectionModule.api$2;->invoke()Lorg/openapitools/client/apis/VideoCallsApi;
 Lio/getstream/video/android/core/internal/module/CoordinatorAuthInterceptor;
 HSPLio/getstream/video/android/core/internal/module/CoordinatorAuthInterceptor;-><clinit>()V
 PLio/getstream/video/android/core/internal/module/CoordinatorAuthInterceptor;-><clinit>()V
