--- stream-video-android-core/src/main/kotlin/org/openapitools/client/apis/LivestreamingApi.kt
@@ -1,106 +1,0 @@
-/*
- * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
- *
- * Licensed under the Stream License;
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.openapitools.client.apis
-
-
-import retrofit2.http.POST
-import retrofit2.http.Path
-
-
-
-
-
-
-import org.openapitools.client.models.StartBroadcastingResponse
-import org.openapitools.client.models.StopBroadcastingResponse
-import org.openapitools.client.models.StopLiveResponse
-import org.openapitools.client.models.StopRecordingResponse
-
-interface LivestreamingApi {
-    /**
-     * Start broadcasting
-     * Starts broadcasting  Required permissions: - StartBroadcasting
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @return [StartBroadcastingResponse]
-     */
-    @POST("/video/call/{type}/{id}/start_broadcasting")
-    suspend fun startBroadcasting(
-        @Path("type") type: String,
-        @Path("id") id: String
-    ): StartBroadcastingResponse
-
-    /**
-     * Stop broadcasting
-     * Stops broadcasting  Required permissions: - StopBroadcasting
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @return [StopBroadcastingResponse]
-     */
-    @POST("/video/call/{type}/{id}/stop_broadcasting")
-    suspend fun stopBroadcasting(
-        @Path("type") type: String,
-        @Path("id") id: String
-    ): StopBroadcastingResponse
-
-    /**
-     * Set call as not live
-     *   Sends events: - call.updated  Required permissions: - UpdateCall
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @return [StopLiveResponse]
-     */
-    @POST("/video/call/{type}/{id}/stop_live")
-    suspend fun stopLive(
-        @Path("type") type: String,
-        @Path("id") id: String
-    ): StopLiveResponse
-
-    /**
-     * Stop recording
-     * Stops recording  Sends events: - call.recording_stopped  Required permissions: - StopRecording
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @return [StopRecordingResponse]
-     */
-    @POST("/video/call/{type}/{id}/stop_recording")
-    suspend fun stopRecording(
-        @Path("type") type: String,
-        @Path("id") id: String
-    ): StopRecordingResponse
-
-}
