--- stream-video-android-core/src/main/kotlin/org/openapitools/client/apis/ModerationApi.kt
@@ -1,141 +1,0 @@
-/*
- * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
- *
- * Licensed under the Stream License;
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.openapitools.client.apis
-
-
-import retrofit2.http.Body
-import retrofit2.http.POST
-import retrofit2.http.Path
-
-
-
-
-
-
-import org.openapitools.client.models.BlockUserRequest
-import org.openapitools.client.models.BlockUserResponse
-import org.openapitools.client.models.MuteUsersRequest
-import org.openapitools.client.models.MuteUsersResponse
-import org.openapitools.client.models.RequestPermissionRequest
-import org.openapitools.client.models.RequestPermissionResponse
-import org.openapitools.client.models.UnblockUserRequest
-import org.openapitools.client.models.UnblockUserResponse
-import org.openapitools.client.models.UpdateUserPermissionsRequest
-import org.openapitools.client.models.UpdateUserPermissionsResponse
-
-interface ModerationApi {
-    /**
-     * Block user on a call
-     * Block a user, preventing them from joining the call until they are unblocked.  Sends events: - call.blocked_user  Required permissions: - BlockUser
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @param blockUserRequest
-     * @return [BlockUserResponse]
-     */
-    @POST("/video/call/{type}/{id}/block")
-    suspend fun blockUser(
-        @Path("type") type: String,
-        @Path("id") id: String,
-        @Body blockUserRequest: BlockUserRequest
-    ): BlockUserResponse
-
-    /**
-     * Mute users
-     * Mutes users in a call  Required permissions: - MuteUsers
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @param muteUsersRequest
-     * @return [MuteUsersResponse]
-     */
-    @POST("/video/call/{type}/{id}/mute_users")
-    suspend fun muteUsers(
-        @Path("type") type: String,
-        @Path("id") id: String,
-        @Body muteUsersRequest: MuteUsersRequest
-    ): MuteUsersResponse
-
-    /**
-     * Request permission
-     * Request permission to perform an action  Sends events: - call.permission_request
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @param requestPermissionRequest
-     * @return [RequestPermissionResponse]
-     */
-    @POST("/video/call/{type}/{id}/request_permission")
-    suspend fun requestPermission(
-        @Path("type") type: String,
-        @Path("id") id: String,
-        @Body requestPermissionRequest: RequestPermissionRequest
-    ): RequestPermissionResponse
-
-    /**
-     * Unblocks user on a call
-     * Removes the block for a user on a call. The user will be able to join the call again.  Sends events: - call.unblocked_user  Required permissions: - BlockUser
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @param unblockUserRequest
-     * @return [UnblockUserResponse]
-     */
-    @POST("/video/call/{type}/{id}/unblock")
-    suspend fun unblockUser(
-        @Path("type") type: String,
-        @Path("id") id: String,
-        @Body unblockUserRequest: UnblockUserRequest
-    ): UnblockUserResponse
-
-    /**
-     * Update user permissions
-     * Updates user permissions  Sends events: - call.permissions_updated  Required permissions: - UpdateCallPermissions
-     * Responses:
-     *  - 201: Successful response
-     *  - 400: Bad request
-     *  - 429: Too many requests
-     *
-     * @param type
-     * @param id
-     * @param updateUserPermissionsRequest
-     * @return [UpdateUserPermissionsResponse]
-     */
-    @POST("/video/call/{type}/{id}/user_permissions")
-    suspend fun updateUserPermissions(
-        @Path("type") type: String,
-        @Path("id") id: String,
-        @Body updateUserPermissionsRequest: UpdateUserPermissionsRequest
-    ): UpdateUserPermissionsResponse
-
-}
