+++ stream-video-android-core/src/main/kotlin/org/openapitools/client/models/APIError.kt
@@ -50,7 +50,7 @@ data class APIError (
 
     /* API error code */
     @Json(name = "code")
-    val code: APIError.Code,
+    val code: kotlin.Int,
 
     /* Additional error-specific information */
     @Json(name = "details")
@@ -73,47 +73,3 @@ data class APIError (
     val exceptionFields: kotlin.collections.Map<kotlin.String, kotlin.String>? = null
 
 )
-
-{
-
-    /**
-     * API error code
-     *
-     * Values: internalError,accessKeyError,inputError,authFailed,duplicateUsername,rateLimited,notFound,notAllowed,eventNotSupported,channelFeatureNotSupported,messageTooLong,multipleNestingLevel,payloadTooBig,expiredToken,tokenNotValidYet,tokenUsedBeforeIat,invalidTokenSignature,customCommandEndpointMissing,customCommandEndpointEqualCallError,connectionIdNotFound,coolDown,queryChannelPermissionsMismatch,tooManyConnections,notSupportedInPushV1,moderationFailed,videoProviderNotConfigured,videoInvalidCallId,videoCreateCallFailed,appSuspended,videoNoDatacentersAvailable,videoJoinCallFailure,queryCallsPermissionsMismatch
-     */
-    enum class Code(val value: kotlin.String) {
-        @Json(name = "internal-error") internalError("internal-error"),
-        @Json(name = "access-key-error") accessKeyError("access-key-error"),
-        @Json(name = "input-error") inputError("input-error"),
-        @Json(name = "auth-failed") authFailed("auth-failed"),
-        @Json(name = "duplicate-username") duplicateUsername("duplicate-username"),
-        @Json(name = "rate-limited") rateLimited("rate-limited"),
-        @Json(name = "not-found") notFound("not-found"),
-        @Json(name = "not-allowed") notAllowed("not-allowed"),
-        @Json(name = "event-not-supported") eventNotSupported("event-not-supported"),
-        @Json(name = "channel-feature-not-supported") channelFeatureNotSupported("channel-feature-not-supported"),
-        @Json(name = "message-too-long") messageTooLong("message-too-long"),
-        @Json(name = "multiple-nesting-level") multipleNestingLevel("multiple-nesting-level"),
-        @Json(name = "payload-too-big") payloadTooBig("payload-too-big"),
-        @Json(name = "expired-token") expiredToken("expired-token"),
-        @Json(name = "token-not-valid-yet") tokenNotValidYet("token-not-valid-yet"),
-        @Json(name = "token-used-before-iat") tokenUsedBeforeIat("token-used-before-iat"),
-        @Json(name = "invalid-token-signature") invalidTokenSignature("invalid-token-signature"),
-        @Json(name = "custom-command-endpoint-missing") customCommandEndpointMissing("custom-command-endpoint-missing"),
-        @Json(name = "custom-command-endpoint=call-error") customCommandEndpointEqualCallError("custom-command-endpoint=call-error"),
-        @Json(name = "connection-id-not-found") connectionIdNotFound("connection-id-not-found"),
-        @Json(name = "cool-down") coolDown("cool-down"),
-        @Json(name = "query-channel-permissions-mismatch") queryChannelPermissionsMismatch("query-channel-permissions-mismatch"),
-        @Json(name = "too-many-connections") tooManyConnections("too-many-connections"),
-        @Json(name = "not-supported-in-push-v1") notSupportedInPushV1("not-supported-in-push-v1"),
-        @Json(name = "moderation-failed") moderationFailed("moderation-failed"),
-        @Json(name = "video-provider-not-configured") videoProviderNotConfigured("video-provider-not-configured"),
-        @Json(name = "video-invalid-call-id") videoInvalidCallId("video-invalid-call-id"),
-        @Json(name = "video-create-call-failed") videoCreateCallFailed("video-create-call-failed"),
-        @Json(name = "app-suspended") appSuspended("app-suspended"),
-        @Json(name = "video-no-datacenters-available") videoNoDatacentersAvailable("video-no-datacenters-available"),
-        @Json(name = "video-join-call-failure") videoJoinCallFailure("video-join-call-failure"),
-        @Json(name = "query-calls-permissions-mismatch") queryCallsPermissionsMismatch("query-calls-permissions-mismatch");
-    }
-
-}
