--- stream-video-android-core/src/main/kotlin/org/openapitools/client/models/APNSRequest.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
- *
- * Licensed under the Stream License;
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress(
-    "ArrayInDataClass",
-    "EnumEntryName",
-    "RemoveRedundantQualifierName",
-    "UnusedImport"
-)
-
-package org.openapitools.client.models
-
-
-
-
-
-import com.squareup.moshi.Json
-
-/**
- *
- *
- * @param body
- * @param title
- */
-
-
-data class APNSRequest (
-
-    @Json(name = "body")
-    val body: kotlin.String? = null,
-
-    @Json(name = "title")
-    val title: kotlin.String? = null
-
-)
